+++ src/components/UI/Column/Column.tsx
@@ -8,6 +8,7 @@ import { Droppable, Draggable } from 'react-beautiful-dnd';
 import { BoardSlice } from '../../../store/reducers/BoardReducer';
 import { useAppDispatch, useAppSelector } from '../../../hooks/redux';
 import ColumnHeader from './ColumnHeader/ColumnHeader';
+import { CreateBoardModalForm } from '../../../types/modalType';
 
 interface IColumnProps {
   column: IColumn;
@@ -16,9 +17,10 @@ interface IColumnProps {
 
 const Column: FC<IColumnProps> = ({ column, refetchAdd }) => {
   const { title, _id: columnId, order, boardId } = column;
-  const [deleteColumn, {}] = useDeleteColumnMutation();
+
+  const [deleteColumn, { isLoading: isDeleting }] = useDeleteColumnMutation();
+  const [createNewTask, { isLoading: isCreating }] = useCreateNewTaskMutation();
   const { data, refetch } = useGetAllTasksQuery({ boardId, columnId });
-  const [createNewTask, {}] = useCreateNewTaskMutation();
   const { setLocalColumnTasks } = BoardSlice.actions;
   const dispatch = useAppDispatch();
   const { columnsTasks } = useAppSelector((state) => state.BoardReducer);
@@ -33,24 +35,63 @@ const Column: FC<IColumnProps> = ({ column, refetchAdd }) => {
     if (data && !columnsTasks.get(columnId)) {
       dispatch(setLocalColumnTasks([columnId, [...(data as ITask[])]]));
     }
-  }, [data]);
+  }, [columnId, columnsTasks, data, dispatch, setLocalColumnTasks]);
+  //Modal manipulations
+  ////////////////////////
+  const [isModalDeleteOpen, setModalDeleteOpen] = useState(false);
+  useEffect(() => {
+    if (!isDeleting) {
+      setModalDeleteOpen(false);
+    }
+  }, [isDeleting]);
+  const [isModalCreateTaskOpen, setModalCreateTaskOpen] = useState(false);
+  useEffect(() => {
+    if (!isCreating) {
+      setModalCreateTaskOpen(false);
+    }
+  }, [isCreating]);
+  ///////////////////////////////////////////////////
 
   const deleteColumnOnClick = () => {
-    deleteColumn({ boardId, columnId });
+    setModalDeleteOpen(true);
   };
 
-  const createNewTaskOnClick = async () => {
+  function confirmDeleteColumn(e: React.MouseEvent<HTMLElement>) {
+    const target = (e.target as HTMLElement).closest('input');
+    const value = target?.value;
+    if (value == 'No' || value == 'Нет') {
+      setModalDeleteOpen(false);
+    }
+    if (value == 'Yes' || value == 'Да') {
+      deleteColumn({ boardId, columnId });
+    }
+  }
+
+  const createNewTaskOnClick = () => {
+    setModalCreateTaskOpen(true);
+  };
+  function cancelTaskHandler(e: React.MouseEvent<HTMLInputElement>) {
+    const input = (e.target as HTMLElement).closest('input');
+    const value = input?.value;
+    if (value && (value == 'Cancel' || value == 'Отмена')) {
+      setModalCreateTaskOpen(false);
+    }
+  }
+
+  async function submitCreateTaskHandler(formData: CreateBoardModalForm) {
+    const { title } = formData;
     const body = {
-      title: Date.now(),
+      title: title,
       order: (data as []).length,
       description: 'string',
-      userId: 0,
+      userId: localStorage.getItem('id'),
       users: ['string'],
     };
     const newTask = await createNewTask({ boardId, columnId, body }).unwrap();
     const newColumnTasks = [...(columnsTasks.get(columnId) as ITask[]), newTask];
     dispatch(setLocalColumnTasks([columnId, newColumnTasks as ITask[]]));
-  };
+  }
+
   return (
     <Draggable draggableId={columnId} index={order}>
       {(provided) => (
@@ -65,6 +106,13 @@ const Column: FC<IColumnProps> = ({ column, refetchAdd }) => {
             deleteColumnOnClick={deleteColumnOnClick}
             createNewTaskOnClick={createNewTaskOnClick}
             dragHandleProps={provided.dragHandleProps}
+            isModalDeleteOpen={isModalDeleteOpen}
+            isDeleting={isDeleting}
+            confirmDeleteColumn={confirmDeleteColumn}
+            isCreating={isCreating}
+            isModalCreateTaskOpen={isModalCreateTaskOpen}
+            cancelTaskHandler={cancelTaskHandler}
+            submitCreateTaskHandler={submitCreateTaskHandler}
           />
           <Droppable droppableId={columnId} type="task">
             {(provided) => (
