+++ src/components/UI/Task/Task.tsx
@@ -1,9 +1,11 @@
-import React, { FC } from 'react';
+import React, { FC, useEffect, useState } from 'react';
 import { ITask } from '../../../types/boardTypes';
 import cl from './Task.module.scss';
 import { useDeleteTaskMutation } from '../../../API/tasksCalls';
 import { Draggable } from 'react-beautiful-dnd';
-import { useTranslate } from '../../../hooks/useTranslate';
+import { Modal } from '../../Modal/modal';
+import ConfirmModal from '../../Modal/modals/confirmModal';
+import removeIcon from '../../../assets/removeTask.svg';
 import { useAppDispatch, useAppSelector } from '../../../hooks/redux';
 import { BoardSlice } from '../../../store/reducers/BoardReducer';
 
@@ -16,18 +18,35 @@ interface ITaskProps {
 
 const Task: FC<ITaskProps> = ({ task, boardId, columnId, index }) => {
   const { title, _id: id } = task;
-  const [deleteTask, {}] = useDeleteTaskMutation();
-  const T = useTranslate();
+  const [deleteTask, { isLoading }] = useDeleteTaskMutation();
+  //Modal manipulations
+  ////////////////////////
+  const [isModalOpen, setModalOpen] = useState(false);
+  useEffect(() => {
+    if (!isLoading) {
+      setModalOpen(false);
+    }
+  }, [isLoading]);
+  ///////////////////////////////////////////////////
   const { columnsTasks } = useAppSelector((state) => state.BoardReducer);
   const dispatch = useAppDispatch();
   const { setLocalColumnTasks } = BoardSlice.actions;
-
   const deleteTaskOnClick = () => {
-    deleteTask({ boardId, columnId, id });
-    const newColumnTasks = columnsTasks.get(columnId)?.filter((task) => task._id !== id);
-    dispatch(setLocalColumnTasks([columnId, newColumnTasks as ITask[]]));
-    console.log(columnsTasks.get(columnId));
+    setModalOpen(true);
   };
+
+  function confirmDeleteTask(e: React.MouseEvent<HTMLElement>) {
+    const target = (e.target as HTMLElement).closest('input');
+    const value = target?.value;
+    if (value == 'No' || value == 'Нет') {
+      setModalOpen(false);
+    }
+    if (value == 'Yes' || value == 'Да') {
+      deleteTask({ boardId, columnId, id });
+      const newColumnTasks = columnsTasks.get(columnId)?.filter((task) => task._id !== id);
+      dispatch(setLocalColumnTasks([columnId, newColumnTasks as ITask[]]));
+    }
+  }
 
   return (
     <Draggable draggableId={id} index={index}>
@@ -39,7 +58,14 @@ const Task: FC<ITaskProps> = ({ task, boardId, columnId, index }) => {
           ref={provided.innerRef}
         >
           <h3>{title}</h3>
-          <button onClick={deleteTaskOnClick}>{T('Task.deleteTask')}</button>
+          <button className={cl.removeBtn} onClick={deleteTaskOnClick}>
+            <img className={cl.icon} src={removeIcon} alt="Benefits icon"></img>
+          </button>
+          {isModalOpen && (
+            <Modal>
+              <ConfirmModal handler={confirmDeleteTask} isLoading={isLoading}></ConfirmModal>
+            </Modal>
+          )}
         </div>
       )}
     </Draggable>
